added district mapping to district-tip job
