--- data/district_tip_testdata.txt
@@ -1,7 +1,0 @@
-07290D3599E7A0D62097A346EFCC1FB5,E7750A37CAB07D0DFF0AF7E3573AC141,2013-01-01 00:00:00,2013-01-01 00:02:00,120,0.44,-73.956528,40.716976,-73.962440,40.715008,CSH,3.50,0.50,0.50,1.00,0.00,4.50,Harlem
-22D70BF00EEB0ADC83BA8177BB861991,3FF2709163DE7036FCAA4E5A3324E4BF,2013-01-01 00:02:00,2013-01-01 00:02:00,0,0.00,0.000000,0.000000,0.000000,0.000000,CSH,27.00,0.00,0.50,3.00,0.00,27.50,Brooklyn
-0EC22AAF491A8BD91F279350C2B010FD,778C92B26AE78A9EBDF96B49C67E4007,2013-01-01 00:01:00,2013-01-01 00:03:00,120,0.71,-73.973145,40.752827,-73.965897,40.760445,CSH,4.00,0.50,0.50,4.00,0.00,5.00,Harlem
-1390FB380189DF6BBFDA4DC847CAD14F,BE317B986700F63C43438482792C8654,2013-01-01 00:01:00,2013-01-01 00:03:00,120,0.48,-74.004173,40.720947,-74.003838,40.726189,CSH,4.00,0.50,0.50,2.00,0.00,5.00,Midtown
-3B4129883A1D05BE89F2C929DE136281,7077F9FD5AD649AEACA4746B2537E3FA,2013-01-01 00:01:00,2013-01-01 00:03:00,120,0.61,-73.987373,40.724861,-73.983772,40.730995,CRD,4.00,0.50,0.50,8.00,0.00,5.00,Midtown
-256C6A3CA47F81497C1F5038438B54B2,8FA961982C3FEF4386FB05B4667A2580,2013-01-01 00:00:00,2013-01-01 00:05:00,300,0.13,-73.969841,40.797359,-73.961899,40.812466,CSH,6.50,0.50,0.50,2.00,0.00,7.50,Brooklyn
-655E773C92FA446353D5C8B7416BE818,B6EAE07E2AD023B387EE5F09BB2D89E3,2013-01-01 00:02:00,2013-01-01 00:05:00,180,1.55,-74.003197,40.733032,-74.012985,40.717377,CRD,6.00,0.50,0.50,3.00,0.00,8.62,Brooklyn
